chore: Move the world file into step_definitions
