+++ cucumber.json
@@ -1,7 +1,7 @@
 {
   "default": {
     "loader": ["ts-node/esm"],
-    "import": ["features/world.ts", "features/**/*.ts"],
+    "import": ["features/step_definitions/world.ts", "features/**/*.ts"],
     "tags": "not @wip",
     "order": "random",
     "parallel": 5,
@@ -10,7 +10,7 @@
   },
   "wip": {
     "loader": ["ts-node/esm"],
-    "import": ["features/world.ts", "features/**/*.ts"],
+    "import": ["features/step_definitions/world.ts", "features/**/*.ts"],
     "tags": "@wip",
     "format": ["@cucumber/pretty-formatter"]
   }
